# openapi.json
{"openapi":"3.1.0","info":{"title":"Bund-ZuwendungsGraph API 🕸️","description":"\n**Sovereign Knowledge Source** für den Bundes-Formularschrank.\n\nDiese API ermöglicht den programmatischen Zugriff auf den Knowledge Graph der deutschen Förderrichtlinien. \nSie kombiniert semantische Vektorsuche mit Graph-RAG (Beziehungen zwischen Dokumenten) und bietet \nautomatisierte Regel-Extraktion via LLM.\n\n### Kernfunktionen:\n- **Hybrid Search**: BM25 + Vector Embeddings (IONOS/Mistral).\n- **Graph Traversal**: Auflösung von Dokumenten-Beziehungen (REFERENCES, SUPERSEDES).\n- **Deep Parsing**: Strukturierte Extraktion aus PDFs via Docling.\n- **RAG Answer Engine**: Kontextbezogene Antworten basierend auf offiziellen Richtlinien.\n\n*Entwickelt als Teil der Forschungsinitiative für transparente Zuwendungsprozesse.*\n","contact":{"name":"DigitalAlchemisten Team","url":"https://digitalalchemisten.de/"},"version":"2.3.1"},"paths":{"/api-info":{"get":{"tags":["System"],"summary":"Api Info","description":"Returns API metadata and links.","operationId":"api_info_api_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/search":{"get":{"summary":"Search","description":"Führt eine hybride Suche (Vektor + Graph) durch und generiert optional eine KI-Antwort.","operationId":"search_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","description":"Die Suchanfrage","title":"Q"},"description":"Die Suchanfrage"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","description":"Maximale Anzahl der Ergebnisse","default":5,"title":"Limit"},"description":"Maximale Anzahl der Ergebnisse"},{"name":"ministerium","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter nach Ministerium","title":"Ministerium"},"description":"Filter nach Ministerium"},{"name":"kuerzel","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter nach Kürzel","title":"Kuerzel"},"description":"Filter nach Kürzel"},{"name":"stand_after","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter nach Datum (Stand nach)","title":"Stand After"},"description":"Filter nach Datum (Stand nach)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/advanced":{"get":{"summary":"Search Advanced","description":"PHASE 1 GRAPH RAG: Advanced Hybrid Search with BM25 + RRF + Reranking.\n\nPipeline:\n1. Multi-Retrieval: BM25 (sparse) + Vector (dense)\n2. RRF Fusion: Reciprocal Rank Fusion\n3. Cross-Encoder Reranking: Semantic reranking for German legal text\n4. Graph Expansion: Multi-hop context (REFERENCES, SUPERSEDES)\n5. Answer Generation: LLM-based answer with provenance\n\nReturns:\n    {\n        \"answer\": str (if generate_answer=true),\n        \"results\": List[Dict],\n        \"metadata\": {\n            \"retrieval_strategy\": \"bm25+vector+reranking\",\n            \"num_results\": int,\n            \"features_enabled\": {...}\n        }\n    }","operationId":"search_advanced_search_advanced_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","description":"Die Suchanfrage","title":"Q"},"description":"Die Suchanfrage"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","description":"Maximale Anzahl der Ergebnisse","default":5,"title":"Limit"},"description":"Maximale Anzahl der Ergebnisse"},{"name":"ministerium","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter nach Ministerium","title":"Ministerium"},"description":"Filter nach Ministerium"},{"name":"kuerzel","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter nach Kürzel","title":"Kuerzel"},"description":"Filter nach Kürzel"},{"name":"stand_after","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter nach Datum (Stand nach)","title":"Stand After"},"description":"Filter nach Datum (Stand nach)"},{"name":"use_bm25","in":"query","required":false,"schema":{"type":"boolean","description":"BM25 Sparse Retrieval aktivieren","default":true,"title":"Use Bm25"},"description":"BM25 Sparse Retrieval aktivieren"},{"name":"use_reranking","in":"query","required":false,"schema":{"type":"boolean","description":"Cross-Encoder Reranking aktivieren","default":true,"title":"Use Reranking"},"description":"Cross-Encoder Reranking aktivieren"},{"name":"use_query_enhancement","in":"query","required":false,"schema":{"type":"boolean","description":"LLM-basierte Query-Optimierung (HyDE) aktivieren (langsam!)","default":false,"title":"Use Query Enhancement"},"description":"LLM-basierte Query-Optimierung (HyDE) aktivieren (langsam!)"},{"name":"multi_hop","in":"query","required":false,"schema":{"type":"boolean","description":"Multi-Hop Graph Traversal aktivieren","default":true,"title":"Multi Hop"},"description":"Multi-Hop Graph Traversal aktivieren"},{"name":"generate_answer","in":"query","required":false,"schema":{"type":"boolean","description":"KI-Antwort generieren","default":true,"title":"Generate Answer"},"description":"KI-Antwort generieren"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graph/expand-context":{"post":{"tags":["Graph Logic"],"summary":"Expand Context Endpoint","description":"**Context-Aware Compliance Mapping**\n\nAnalysiert übergebene Text-Chunks (z.B. aus einer neuen Förderrichtlinie) und expandiert \ndiese gegen den Knowledge Graph.\n\nProzess:\n1. **Citation Matching**: Findet harte Referenzen (z.B. \"NKBF 98\").\n2. **Concept Expansion**: Findet implizite Regeln für erkannte Konzepte (z.B. \"Reisekosten\" -> \"BRKG\").\n\nReturns:\n    Ein strukturiertes Regel-Paket (`mapped_regulations`), das für den Prüfagenten optimiert ist.","operationId":"expand_context_endpoint_graph_expand_context_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpandContextRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpandContextResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health-raw":{"get":{"summary":"Health Raw","description":"Health check endpoint.","operationId":"health_raw_health_raw_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ExpandContextRequest":{"properties":{"context_label":{"type":"string","title":"Context Label","description":"Label for the context, e.g. 'Guidelines_Drone_2024'"},"text_chunks":{"items":{"type":"string"},"type":"array","title":"Text Chunks","description":"List of text chunks from the guideline"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata","description":"Additional metadata like agency, year"}},"type":"object","required":["context_label","text_chunks"],"title":"ExpandContextRequest"},"ExpandContextResponse":{"properties":{"compliance_context_id":{"type":"string","title":"Compliance Context Id"},"mapped_regulations":{"items":{"$ref":"#/components/schemas/MappedRegulation"},"type":"array","title":"Mapped Regulations"}},"type":"object","required":["compliance_context_id","mapped_regulations"],"title":"ExpandContextResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MappedRegulation":{"properties":{"category":{"type":"string","title":"Category"},"source_doc":{"type":"string","title":"Source Doc"},"rules":{"items":{"$ref":"#/components/schemas/MappedRule"},"type":"array","title":"Rules"}},"type":"object","required":["category","source_doc","rules"],"title":"MappedRegulation"},"MappedRule":{"properties":{"rule_id":{"type":"string","title":"Rule Id"},"content":{"type":"string","title":"Content"},"relevance_reason":{"type":"string","title":"Relevance Reason"}},"type":"object","required":["rule_id","content","relevance_reason"],"title":"MappedRule"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}